+++ rapidfuzz/distance/OSA.hpp
@@ -202,16 +202,16 @@ private:
         detail::Range scores_(scores, scores + score_count);
         if constexpr (MaxLen == 8)
             detail::osa_hyrroe2003_simd<uint8_t>(scores_, PM, str_lens, s2,
-                                                 static_cast<size_t>(score_cutoff));
+                                                 score_cutoff);
         else if constexpr (MaxLen == 16)
             detail::osa_hyrroe2003_simd<uint16_t>(scores_, PM, str_lens, s2,
-                                                  static_cast<size_t>(score_cutoff));
+                                                  score_cutoff);
         else if constexpr (MaxLen == 32)
             detail::osa_hyrroe2003_simd<uint32_t>(scores_, PM, str_lens, s2,
-                                                  static_cast<size_t>(score_cutoff));
+                                                  score_cutoff);
         else if constexpr (MaxLen == 64)
             detail::osa_hyrroe2003_simd<uint64_t>(scores_, PM, str_lens, s2,
-                                                  static_cast<size_t>(score_cutoff));
+                                                  score_cutoff);
     }
 
     template <typename InputIt2>
